requirements.txt
coverage==4.5.1
google-api-core==1.7.0
google-auth-oauthlib==0.3.0
google-auth==1.5.1
google-cloud-bigquery==1.11.2
google-cloud-core==0.29.1
google-cloud-storage==1.13.2
google-cloud==0.34.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
graphviz==0.10.1
matplotlib==2.1.2
mysql-connector-python==8.0.16
mysql-connector==2.2.9
networkx==2.1
pandas-gbq==0.10.0
pandas==0.22.0
psycopg2-binary==2.8.2
pybase64==0.5.0
pydata-google-auth==0.1.3
pydot==1.4.1
pytest-cov==2.6.0
pytest==3.3.2
